test(app): address review comments on #4852
- Add missing ``__init__.py`` to ``tests/unit/app/runner/``,
  ``tests/unit/app/routers/``, and ``tests/contract/runner/`` to match
  the convention used by every other test package in the repo (e.g.
  ``tests/unit/agents/tools/``, ``tests/unit/security/tool_guard/``).
- Move the messages-only ``app`` / ``client`` fixtures out of
  ``tests/unit/app/routers/conftest.py`` and into
  ``test_messages_router.py``. The conftest now only owns the truly
  shared mocks (``workspace_mock``, ``manager_mock``), preventing a
  future ``test_xxx_router.py`` author from silently inheriting a
  messages-only app and debugging 404s.

No behaviour change. All 153 Phase 3 tests still pass.
